+++ src/socket.h
@@ -320,7 +320,7 @@ namespace abc {
 	inline void _client_socket<Log>::send(const void* buffer, std::size_t size, socket::address* address) {
 		Log* log_local = base::log();
 		if (log_local != nullptr) {
-			log_local->put_any(category::abc::socket, severity::abc::debug, 0x10016, "_client_socket::send() >>> size=%lu", (unsigned long)size);
+			log_local->put_any(category::abc::socket, severity::abc::debug, 0x10016, "_client_socket::send() >>> size=%zu", size);
 		}
 
 		if (!base::is_open()) {
@@ -348,7 +348,7 @@ namespace abc {
 
 		if (log_local != nullptr) {
 			log_local->put_binary(category::abc::socket, severity::abc::optional, 0x10066, buffer, size);
-			log_local->put_any(category::abc::socket, severity::abc::optional, 0x1001b, "_client_socket::send() <<< size=%lu", (unsigned long)size);
+			log_local->put_any(category::abc::socket, severity::abc::optional, 0x1001b, "_client_socket::send() <<< size=%zu", size);
 		}
 	}
 
@@ -357,7 +357,7 @@ namespace abc {
 	inline void _client_socket<Log>::receive(void* buffer, std::size_t size, socket::address* address) {
 		Log* log_local = base::log();
 		if (log_local != nullptr) {
-			log_local->put_any(category::abc::socket, severity::abc::debug, 0x1001c, "_client_socket::receive() >>> size=%lu", (unsigned long)size);
+			log_local->put_any(category::abc::socket, severity::abc::debug, 0x1001c, "_client_socket::receive() >>> size=%zu", size);
 		}
 
 		if (!base::is_open()) {
@@ -385,7 +385,7 @@ namespace abc {
 
 		if (log_local != nullptr) {
 			log_local->put_binary(category::abc::socket, severity::abc::optional, 0x10067, buffer, size);
-			log_local->put_any(category::abc::socket, severity::abc::optional, 0x10021, "_client_socket::receive() <<< size=%lu", (unsigned long)size);
+			log_local->put_any(category::abc::socket, severity::abc::optional, 0x10021, "_client_socket::receive() <<< size=%zu", size);
 		}
 	}
